+++ src/main/java/net/soulsweaponry/items/abilities/use/SoulReleasePowerBased.java
@@ -1,10 +1,12 @@
 package net.soulsweaponry.items.abilities.use;
 
+import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.LivingEntity;
+import net.minecraft.entity.attribute.EntityAttributes;
+import net.minecraft.entity.passive.TameableEntity;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemStack;
-import net.minecraft.server.world.ServerWorld;
 import net.minecraft.sound.SoundCategory;
 import net.minecraft.text.Text;
 import net.minecraft.util.Formatting;
@@ -12,70 +14,115 @@ import net.minecraft.util.Hand;
 import net.minecraft.util.TypedActionResult;
 import net.minecraft.util.math.Vec3d;
 import net.minecraft.world.World;
-import net.soulsweaponry.config.ConfigConstructor;
-import net.soulsweaponry.entity.mobs.Forlorn;
-import net.soulsweaponry.entity.mobs.SoulReaperGhost;
-import net.soulsweaponry.entity.mobs.Soulmass;
 import net.soulsweaponry.items.abilities.ISummonAlliesAbility;
 import net.soulsweaponry.items.abilities.targetdeath.ISoulHarvest;
 import net.soulsweaponry.particles.ParticleEvents;
 import net.soulsweaponry.particles.ParticleHandler;
-import net.soulsweaponry.registry.EntityRegistry;
 import net.soulsweaponry.registry.SoundRegistry;
+import net.soulsweaponry.util.WeaponUtil;
+import org.jetbrains.annotations.Nullable;
 
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
-public record SoulReleasePowerBased(int maxSummons, Map<Integer, EntityType<?>> entityPowerMap) implements ISoulHarvest, ISummonAlliesAbility {
-
-    private static final SoulReleasePowerBased SOUL_RELEASE = new SoulReleasePowerBased(
-            (int) ConfigConstructor.soul_reaper_summoned_allies_cap,
-            Map.of(
-                    3, EntityRegistry.SOUL_REAPER_GHOST,
-                    10, EntityRegistry.FORLORN,
-                    30, EntityRegistry.SOULMASS
-            )
-    );
+/**
+ * Use to spawn entities based on the amount of souls the item has harvested by fetching {@link #getSouls(ItemStack)}.
+ * The ability will prioritize the best entity based on power needed.
+ * <p>
+ * For example: if the required for 3 entities is 5, 10 and 50, and the power is 67, first entity is the one requiring 50 and
+ * so 50 souls is deducted. 17 souls remain, so then it would choose the one requiring 10. 7 souls remain, and so entity
+ * needing 5 is chosen. 2 souls remain so nothing happens.
+ * <p>
+ * If the player sneaks, then the {@code second best} entity will always be chosen.
+ * <p>
+ * Being in creative will always spawn the best entity.
+ *
+ * @param maxSummons max amount of allies that can exist at once
+ * @param entityPowerMap tree map containing the mappings of the entity and how many souls that are required
+ * @param bonusHealthPerPower bonus health per power (per soul)
+ * @param bonusHealthIncreasePerLvl added to {@code bonusHealthPerPower} before the bonus health per soul calculation, is based on weapon level
+ * @param maxBonusHealth max bonus health the summon can get
+ * @param bonusAttackPerPower bonus attack damage per power (per soul)
+ * @param bonusAttackIncreasePerLvl added to {@code bonusAttackPerPower} before the bonus attack per soul calculation, is based on weapon level
+ * @param maxBonusAttack max bonus attack damage the summon can get
+ */
+public record SoulReleasePowerBased(int maxSummons, NavigableMap<Integer, EntityType<?>> entityPowerMap,
+                                    float bonusHealthPerPower, float bonusHealthIncreasePerLvl, float maxBonusHealth,
+                                    float bonusAttackPerPower, float bonusAttackIncreasePerLvl, float maxBonusAttack
+) implements ISoulHarvest, ISummonAlliesAbility {
 
     @Override
     public TypedActionResult<ItemStack> use(World world, PlayerEntity player, Hand hand, ItemStack stack) {
-        //TODO rewrite the whole thing so that it chooses off of the map based on power
-        // TODO also make it choose second best when sneaking
-        int power = this.getSouls(stack);
-        if (player.isCreative()) power = player.getRandom().nextBetween(5, 50);
-        if (power != 0) {
-            if (power >= 3 && !world.isClient && this.canSummonEntity((ServerWorld) world, player, this.getSummonsListId())) {
-                Vec3d vecBlocksAway = player.getRotationVector().multiply(3).add(player.getPos());
-                ParticleHandler.particleOutburstMap(world, 50, vecBlocksAway.getX(), vecBlocksAway.getY(), vecBlocksAway.getZ(), ParticleEvents.CONJURE_ENTITY_MAP, 1f);
-                world.playSound(null, player.getBlockPos(), SoundRegistry.NIGHTFALL_SPAWN_EVENT, SoundCategory.PLAYERS, 0.8f, 1f);
-                if (power < 10) {
-                    SoulReaperGhost entity = new SoulReaperGhost(EntityRegistry.SOUL_REAPER_GHOST, world);
-                    entity.setPos(vecBlocksAway.x, player.getY() + .1f, vecBlocksAway.z);
-                    entity.setOwner(player);
-                    world.spawnEntity(entity);
-                    this.saveSummonUuid(player, entity.getUuid());
-                    if (!player.isCreative()) this.addAmount(stack, -3);
-                } else if (player.isSneaking() || power < 30) {
-                    Forlorn entity = new Forlorn(EntityRegistry.FORLORN, world);
-                    entity.setPos(vecBlocksAway.x, player.getY() + .1f, vecBlocksAway.z);
-                    entity.setOwner(player);
-                    world.spawnEntity(entity);
-                    this.saveSummonUuid(player, entity.getUuid());
-                    if (!player.isCreative()) this.addAmount(stack, -10);
-                } else {
-                    Soulmass entity = new Soulmass(EntityRegistry.SOULMASS, world);
-                    entity.setPos(vecBlocksAway.x, player.getY() + .1f, vecBlocksAway.z);
-                    entity.setOwner(player);
-                    world.spawnEntity(entity);
-                    this.saveSummonUuid(player, entity.getUuid());
-                    if (!player.isCreative()) this.addAmount(stack, -30);
-                }
-
-                stack.damage(3, player, LivingEntity.getSlotForHand(hand));
-                return TypedActionResult.success(stack, true);
+        int power = player.isCreative() ? this.entityPowerMap.lastKey() : this.getSouls(stack);
+        if (entityPowerMap.isEmpty() || power < entityPowerMap.firstKey()) {
+            return TypedActionResult.fail(stack);
+        }
+        boolean sneaking = player.isSneaking();
+        Map.Entry<Integer, EntityType<?>> chosen = sneaking ? secondBestEntry(power) : bestEntry(power);
+        if (chosen == null) {
+            return TypedActionResult.fail(stack);
+        }
+        EntityType<?> type = chosen.getValue();
+        int cost = chosen.getKey();
+        Vec3d vecBlocksAway = player.getRotationVector().multiply(3).add(player.getPos());
+        ParticleHandler.particleOutburstMap(world, 50, vecBlocksAway.getX(), vecBlocksAway.getY(), vecBlocksAway.getZ(), ParticleEvents.CONJURE_ENTITY_MAP, 1f);
+        world.playSound(null, player.getBlockPos(), SoundRegistry.NIGHTFALL_SPAWN_EVENT, SoundCategory.PLAYERS, 0.8f, 1f);
+        Entity e = type.create(world);
+        if (e != null) {
+            e.setPos(vecBlocksAway.x, player.getY() + .1f, vecBlocksAway.z);
+            if (e instanceof TameableEntity t) {
+                t.setOwner(player);
+            }
+            this.updateStats(e, power, stack);
+            world.spawnEntity(e);
+            this.saveSummonUuid(player, e.getUuid());
+            if (!player.isCreative()) {
+                this.addAmount(stack, - cost);
             }
         }
-        return TypedActionResult.fail(stack);
+        stack.damage(3, player, LivingEntity.getSlotForHand(hand));
+        return TypedActionResult.success(stack, true);
+    }
+
+    /**
+     * Increases the {@link EntityAttributes#GENERIC_MAX_HEALTH} and {@link EntityAttributes#GENERIC_ATTACK_DAMAGE}
+     * based on {@code power} (souls).
+     */
+    private void updateStats(Entity entity, int power, ItemStack stack) {
+        if (entity instanceof LivingEntity living) {
+            int lvl = WeaponUtil.getUpgradeLevel(stack);
+            var maxHealth = living.getAttributeInstance(EntityAttributes.GENERIC_MAX_HEALTH);
+            if (maxHealth != null) {
+                double newMax = living.getMaxHealth() + (this.bonusHealthPerPower + this.bonusHealthIncreasePerLvl * lvl) * power;
+                newMax = Math.min(this.maxBonusHealth, newMax);
+                maxHealth.setBaseValue(newMax);
+                living.setHealth((float) newMax);
+            }
+            var atk = living.getAttributeInstance(EntityAttributes.GENERIC_ATTACK_DAMAGE);
+            if (atk != null) {
+                double newDmg = atk.getValue() + (this.bonusAttackPerPower + this.bonusAttackIncreasePerLvl * lvl) * power;
+                newDmg = Math.min(this.maxBonusAttack, newDmg);
+                atk.setBaseValue(newDmg);
+            }
+        }
+    }
+
+    /**
+     * Gets the best entry based on power.
+     * Returns null if the power is under the smallest threshold.
+     */
+    @Nullable
+    private Map.Entry<Integer, EntityType<?>> bestEntry(int power) {
+        return entityPowerMap.floorEntry(power);
+    }
+
+    /**
+     * Gets the second-best entry based on power.
+     * Returns null if the power is under the smallest threshold.
+     */
+    @Nullable
+    private Map.Entry<Integer, EntityType<?>> secondBestEntry(int power) {
+        var best = entityPowerMap.floorEntry(power);
+        return (best == null) ? null : Objects.requireNonNullElse(entityPowerMap.lowerEntry(best.getKey()), best);
     }
 
     @Override