+++ src/main/java/net/soulsweaponry/items/scythe/SoulReaper.java
@@ -1,26 +1,46 @@
 package net.soulsweaponry.items.scythe;
 
 import net.minecraft.client.render.item.BuiltinModelItemRenderer;
+import net.minecraft.entity.EntityType;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.ToolMaterial;
 import net.soulsweaponry.client.renderer.item.SoulReaperRenderer;
 import net.soulsweaponry.config.ConfigConstructor;
 import net.soulsweaponry.items.SoulHarvestingItem;
+import net.soulsweaponry.items.abilities.use.SoulReleasePowerBased;
+import net.soulsweaponry.registry.EntityRegistry;
 import software.bernie.geckolib.animatable.GeoItem;
 import software.bernie.geckolib.animatable.client.GeoRenderProvider;
 import software.bernie.geckolib.animatable.instance.AnimatableInstanceCache;
 import software.bernie.geckolib.animation.*;
 import software.bernie.geckolib.util.GeckoLibUtil;
 
+import java.util.NavigableMap;
+import java.util.TreeMap;
 import java.util.function.Consumer;
 
 public class SoulReaper extends SoulHarvestingItem implements GeoItem {
 
     private final AnimatableInstanceCache factory = GeckoLibUtil.createInstanceCache(this);
+    private static final NavigableMap<Integer, EntityType<?>> THRESHOLDS = new TreeMap<>();
+    private static final SoulReleasePowerBased SOUL_RELEASE = new SoulReleasePowerBased(
+            (int) ConfigConstructor.soul_reaper_summoned_allies_cap,
+            THRESHOLDS,
+            ConfigConstructor.soul_reaper_summon_bonus_health_per_soul, ConfigConstructor.soul_reaper_summon_bonus_health_per_soul_addition_per_level,
+            ConfigConstructor.soul_reaper_summon_max_bonus_health,
+            ConfigConstructor.soul_reaper_summon_bonus_attack_damage_per_soul, ConfigConstructor.soul_reaper_summon_bonus_attack_damage_per_soul_addition_per_level,
+            ConfigConstructor.soul_reaper_summon_max_bonus_attack_damage
+    );
+
+    static {
+        THRESHOLDS.put(3, EntityRegistry.SOUL_REAPER_GHOST);
+        THRESHOLDS.put(10, EntityRegistry.FORLORN);
+        THRESHOLDS.put(30, EntityRegistry.SOULMASS);
+    }
 
     public SoulReaper(ToolMaterial toolMaterial, Settings settings) {
         super(toolMaterial, (int) ConfigConstructor.soul_reaper_damage, ConfigConstructor.soul_reaper_attack_speed, settings);
-        //TODO add SoulReleasePowerBased here
+        this.addAbility(SOUL_RELEASE);
     }
 
     private PlayState predicate(AnimationState<?> event){